requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
pandas>=1.3.4,<1.4
requests>=2.26.0,<2.27
ta>=0.7.0,<0.8
yahoofinancials==1.14
drf-spectacular>=0.15.1,<0.16
beautifulsoup4>=4.10.0,<4.11
Pillow>=8.2.0,<8.3.0